requirements.txt
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keras-rl==0.4.2
numpy==1.16.4
python-chess==0.28.0
tensorboard==1.9.0
tensorflow==1.9.0
